requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
Django==5.2
django-import-export==4.3.14
sqlparse==0.5.5
tablib==3.9.0
